Simple isBetween Method With Inclusivity Parameter
This gives users the ability to choose the level of inclusivity the method uses. Implementing inclusivity this way reduces total lines or code, lines of code changed, backwards compatibility and changes to documentation.

Nearly the same as proposed, but eliminates the 'options' parameter. Nothing I've seen in Moment uses that approach.